+++ src/main/net-constraints-nunit3/ValidationConstraints.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Linq;
 using System.Text;
+using System.Xml.Schema;
 using NUnit.Framework.Constraints;
 using Org.XmlUnit.Validation;
 using InputBuilder = Org.XmlUnit.Builder.Input;
@@ -22,7 +23,7 @@ using InputBuilder = Org.XmlUnit.Builder.Input;
 namespace Org.XmlUnit.Constraints {
 
     /// <summary>
-    /// Constraint that validates a document against a given XML
+    /// Constraint that validates a document against a given W3C XML
     /// schema.
     /// </summary>
     public class SchemaValidConstraint : Constraint {
@@ -43,6 +44,22 @@ namespace Org.XmlUnit.Constraints {
             validator.SchemaSources = schema
                 .Select(s => InputBuilder.From(s).Build())
                 .ToArray();
+        }
+
+        /// <summary>
+        /// Creates the constraint validating against the given schema.
+        /// </summary>
+        /// <remarks>
+        ///   <para>
+        ///  since XMLUnit 2.3.0
+        ///   </para>
+        /// </remarks>
+        public SchemaValidConstraint(XmlSchema schema) : base(schema) {
+            if (schema == null) {
+                throw new ArgumentNullException("schema");
+            }
+            validator = Validator.ForLanguage(Languages.W3C_XML_SCHEMA_NS_URI);
+            validator.Schema = schema;
         }
 
         /// <inheritdoc/>
@@ -69,7 +86,11 @@ namespace Org.XmlUnit.Constraints {
 
             /// <inheritdoc/>
             public override void WriteMessageTo(MessageWriter writer) {
-                if (constraint.validator.SchemaSources.Count(s => !string.IsNullOrEmpty(s.SystemId)) > 0) {
+                if (constraint.validator.Schema != null) {
+                    writer.Write("{0} validates against {1}",
+                                 GrabSystemId(ActualValue as ISource) ?? "instance",
+                                 constraint.validator.Schema.SourceUri ?? " the given schema");
+                } else if (constraint.validator.SchemaSources.Count(s => !string.IsNullOrEmpty(s.SystemId)) > 0) {
                     writer.WriteLine("{0} does not validate against {1}",
                         GrabSystemId(ActualValue as ISource) ?? "instance",
                         GrabSystemIds());